+++ src/test/wrapper.cpp
@@ -5,9 +5,10 @@
 #include <stdio.h>
 rust::String optimize(rust::String circ_string, rust::String cost_func,
                       rust::String ecc_path, rust::String gateset,
-                      int timeout) {
-  return optimize_(std::string(circ_string), std::string(cost_func),
-                   std::string(ecc_path), std::string(gateset), timeout);
+                      float timeout)
+{
+    return optimize_(std::string(circ_string), std::string(cost_func),
+                     std::string(ecc_path), std::string(gateset), timeout);
 }
 // int main() {
 //   int result = optimize();
